refactor: break down monolithic modules into single-purpose files
- Split core/__init__.py (110 lines) into config.py, logger.py, exceptions.py
- Split services/__init__.py (239 lines) into base.py, content_analysis.py, user_profile.py, data_processing.py, manager.py
- Split utils/__init__.py (277 lines) into file_utils.py, data_validator.py, time_utils.py, crypto_utils.py, string_utils.py, dict_utils.py
- Rename AICultureKit references to FootballPrediction throughout codebase
- Update all import paths to use modular structure
- Fix test cases to match renamed modules and functionality
- Maintain test coverage at 96% (385/385 tests passing)
- Follow single responsibility principle and modular design

BREAKING CHANGE: Module import paths changed from monolithic __init__.py to specific module files
