+++ src/models/__init__.py
@@ -1,5 +1,5 @@
 """
-AICultureKit 数据模型模块
+足球预测系统数据模型模块
 
 提供系统中使用的数据模型，包括：
 - 用户模型
