+++ src/services/__init__.py
@@ -1,238 +1,27 @@
 """
-AICultureKit 业务服务模块
+足球预测系统业务服务模块
 
 提供系统业务逻辑服务，包括：
 - 内容分析服务
 - 用户画像服务
 - 数据处理服务
-- AI模型调用服务
+- 服务管理器
 """
 
-from abc import ABC, abstractmethod
-from typing import Any, Dict, List, Optional
-
-from src.core import logger
-from src.models import AnalysisResult, Content, User, UserProfile
-
-
-class BaseService(ABC):
-    """基础服务抽象类"""
-
-    def __init__(self, name: str):
-        self.name = name
-        self.logger = logger
-
-    @abstractmethod
-    async def initialize(self) -> bool:
-        """服务初始化"""
-        pass
-
-    @abstractmethod
-    async def shutdown(self) -> None:
-        """服务关闭"""
-        pass
-
-
-class ContentAnalysisService(BaseService):
-    """内容分析服务"""
-
-    def __init__(self):
-        super().__init__("ContentAnalysisService")
-        self._initialized = False
-
-    async def initialize(self) -> bool:
-        """初始化服务"""
-        self.logger.info(f"正在初始化 {self.name}")
-        # TODO: 加载AI模型、连接外部API等
-        self._initialized = True
-        return True
-
-    async def shutdown(self) -> None:
-        """关闭服务"""
-        self.logger.info(f"正在关闭 {self.name}")
-        self._initialized = False
-
-    async def analyze_content(self, content: Content) -> Optional[AnalysisResult]:
-        """分析内容"""
-        if not self._initialized:
-            raise RuntimeError("服务未初始化")
-
-        self.logger.info(f"正在分析内容: {content.id}")
-
-        # TODO: 实现实际的内容分析逻辑
-        # 这里是示例实现
-        analysis_data = {
-            "sentiment": "positive",
-            "keywords": ["文化", "AI", "分析"],
-            "category": "技术",
-            "quality_score": 0.85,
-        }
-
-        return AnalysisResult(
-            id=f"analysis_{content.id}",
-            content_id=content.id,
-            analysis_type="content_analysis",
-            result_data=analysis_data,
-            confidence_score=0.85,
-        )
-
-    async def batch_analyze(self, contents: List[Content]) -> List[AnalysisResult]:
-        """批量分析内容"""
-        results = []
-        for content in contents:
-            result = await self.analyze_content(content)
-            if result:
-                results.append(result)
-        return results
-
-
-class UserProfileService(BaseService):
-    """用户画像服务"""
-
-    def __init__(self) -> None:
-        super().__init__("UserProfileService")
-        self._user_profiles: Dict[str, UserProfile] = {}
-
-    async def initialize(self) -> bool:
-        """初始化服务"""
-        self.logger.info(f"正在初始化 {self.name}")
-        # TODO: 加载用户数据、模型等
-        return True
-
-    async def shutdown(self) -> None:
-        """关闭服务"""
-        self.logger.info(f"正在关闭 {self.name}")
-        self._user_profiles.clear()
-
-    async def generate_profile(self, user: User) -> UserProfile:
-        """生成用户画像"""
-        self.logger.info(f"正在生成用户画像: {user.id}")
-
-        # TODO: 实现实际的用户画像生成逻辑
-        profile = UserProfile(
-            user_id=user.id,
-            interests=["文化", "技术", "AI"],
-            preferences={"content_type": "text", "language": "zh"},
-            behavior_patterns={"active_hours": [9, 10, 11, 14, 15, 16]},
-        )
-
-        self._user_profiles[user.id] = profile
-        return profile
-
-    async def get_profile(self, user_id: str) -> Optional[UserProfile]:
-        """获取用户画像"""
-        return self._user_profiles.get(user_id)
-
-    async def update_profile(
-        self, user_id: str, updates: Dict[str, Any]
-    ) -> Optional[UserProfile]:
-        """更新用户画像"""
-        profile = await self.get_profile(user_id)
-        if not profile:
-            return None
-
-        # 更新画像数据
-        for key, value in updates.items():
-            if hasattr(profile, key):
-                setattr(profile, key, value)
-
-        return profile
-
-
-class DataProcessingService(BaseService):
-    """数据处理服务"""
-
-    def __init__(self):
-        super().__init__("DataProcessingService")
-
-    async def initialize(self) -> bool:
-        """初始化服务"""
-        self.logger.info(f"正在初始化 {self.name}")
-        return True
-
-    async def shutdown(self) -> None:
-        """关闭服务"""
-        self.logger.info(f"正在关闭 {self.name}")
-
-    async def process_text(self, text: str) -> Dict[str, Any]:
-        """处理文本数据"""
-        # TODO: 实现文本处理逻辑（清洗、分词、特征提取等）
-        return {
-            "processed_text": text.strip(),
-            "word_count": len(text.split()),
-            "character_count": len(text),
-        }
-
-    async def process_batch(self, data_list: List[Any]) -> List[Dict[str, Any]]:
-        """批量处理数据"""
-        results = []
-        for data in data_list:
-            if isinstance(data, str):
-                result = await self.process_text(data)
-                results.append(result)
-        return results
-
-
-class ServiceManager:
-    """服务管理器 - 负责统一管理所有业务服务的生命周期和依赖关系"""
-
-    def __init__(self) -> None:
-        self.services: Dict[str, BaseService] = {}
-        self.logger = logger
-
-    def register_service(self, service: BaseService) -> None:
-        """注册服务 - 将服务加入管理器，支持后续统一初始化和管理"""
-        self.services[service.name] = service
-        self.logger.info(f"已注册服务: {service.name}")
-
-    async def initialize_all(self) -> bool:
-        """初始化所有服务 - 按注册顺序依次初始化，任一失败则整体失败"""
-        self.logger.info("正在初始化所有服务...")
-        success = True
-
-        for service in self.services.values():
-            try:
-                # 每个服务独立初始化，失败不影响其他服务的尝试
-                result = await service.initialize()
-                if not result:
-                    success = False
-                    self.logger.error(f"服务初始化失败: {service.name}")
-            except Exception as e:
-                # 捕获异常避免整个初始化流程中断
-                success = False
-                self.logger.error(f"服务初始化异常: {service.name}, {e}")
-
-        return success
-
-    async def shutdown_all(self) -> None:
-        """关闭所有服务 - 确保资源清理，即使某个服务关闭失败也继续处理其他服务"""
-        self.logger.info("正在关闭所有服务...")
-
-        for service in self.services.values():
-            try:
-                await service.shutdown()
-            except Exception as e:
-                # 关闭失败不应阻止其他服务的正常关闭
-                self.logger.error(f"服务关闭异常: {service.name}, {e}")
-
-    def get_service(self, name: str) -> Optional[BaseService]:
-        """获取服务实例 - 提供类型安全的服务访问接口"""
-        return self.services.get(name)
-
-
-# 全局服务管理器实例
-service_manager = ServiceManager()
-
-# 注册默认服务
-service_manager.register_service(ContentAnalysisService())
-service_manager.register_service(UserProfileService())
-service_manager.register_service(DataProcessingService())
+from .base import BaseService
+from .content_analysis import ContentAnalysisService
+from .data_processing import DataProcessingService
+from .manager import ServiceManager, service_manager
+from .user_profile import UserProfileService
 
 __all__ = [
+    # 基础服务
     "BaseService",
+    # 具体服务
     "ContentAnalysisService",
     "UserProfileService",
     "DataProcessingService",
+    # 服务管理
     "ServiceManager",
     "service_manager",
 ]
