+++ src/services/content_analysis.py
@@ -1,0 +1,64 @@
+"""
+足球预测系统内容分析服务模块
+
+提供内容分析和处理功能。
+"""
+
+from typing import List, Optional
+
+from src.models import AnalysisResult, Content
+
+from .base import BaseService
+
+
+class ContentAnalysisService(BaseService):
+    """内容分析服务"""
+
+    def __init__(self):
+        super().__init__("ContentAnalysisService")
+        self._initialized = False
+
+    async def initialize(self) -> bool:
+        """初始化服务"""
+        self.logger.info(f"正在初始化 {self.name}")
+        # TODO: 加载AI模型、连接外部API等
+        self._initialized = True
+        return True
+
+    async def shutdown(self) -> None:
+        """关闭服务"""
+        self.logger.info(f"正在关闭 {self.name}")
+        self._initialized = False
+
+    async def analyze_content(self, content: Content) -> Optional[AnalysisResult]:
+        """分析内容"""
+        if not self._initialized:
+            raise RuntimeError("服务未初始化")
+
+        self.logger.info(f"正在分析内容: {content.id}")
+
+        # TODO: 实现实际的内容分析逻辑
+        # 这里是示例实现
+        analysis_data = {
+            "sentiment": "positive",
+            "keywords": ["足球", "预测", "分析"],
+            "category": "体育",
+            "quality_score": 0.85,
+        }
+
+        return AnalysisResult(
+            id=f"analysis_{content.id}",
+            content_id=content.id,
+            analysis_type="content_analysis",
+            result_data=analysis_data,
+            confidence_score=0.85,
+        )
+
+    async def batch_analyze(self, contents: List[Content]) -> List[AnalysisResult]:
+        """批量分析内容"""
+        results = []
+        for content in contents:
+            result = await self.analyze_content(content)
+            if result:
+                results.append(result)
+        return results
