+++ src/services/data_processing.py
@@ -1,0 +1,43 @@
+"""
+足球预测系统数据处理服务模块
+
+提供数据清洗、处理和特征提取功能。
+"""
+
+from typing import Any, Dict, List
+
+from .base import BaseService
+
+
+class DataProcessingService(BaseService):
+    """数据处理服务"""
+
+    def __init__(self):
+        super().__init__("DataProcessingService")
+
+    async def initialize(self) -> bool:
+        """初始化服务"""
+        self.logger.info(f"正在初始化 {self.name}")
+        return True
+
+    async def shutdown(self) -> None:
+        """关闭服务"""
+        self.logger.info(f"正在关闭 {self.name}")
+
+    async def process_text(self, text: str) -> Dict[str, Any]:
+        """处理文本数据"""
+        # TODO: 实现文本处理逻辑（清洗、分词、特征提取等）
+        return {
+            "processed_text": text.strip(),
+            "word_count": len(text.split()),
+            "character_count": len(text),
+        }
+
+    async def process_batch(self, data_list: List[Any]) -> List[Dict[str, Any]]:
+        """批量处理数据"""
+        results = []
+        for data in data_list:
+            if isinstance(data, str):
+                result = await self.process_text(data)
+                results.append(result)
+        return results
