+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""AICultureKit 测试模块初始化文件"""
+"""足球预测系统测试模块初始化文件"""
